Refactor: 각 컴포넌트 내에 있는 styled-component css 코드 -> styles.tsx파일 생성하여 따로 분리
