+++ src/assets/icons/JosephMovieLogo.tsx
@@ -4,22 +4,18 @@ import { SVGProps } from "react";
 const JosephMovieLogo = (props: SVGProps<SVGSVGElement>) => (
   <svg
     xmlns="http://www.w3.org/2000/svg"
-    width={120}
-    height={60}
-    viewBox="0 0 200 106"
+    width={props.width}
+    height={props.height}
+    viewBox="0 0 200 101"
     {...props}
   >
     <path
       fill="white"
-      d="M19 12.5c0 4.2.3 5.5 1.5 5.5S22 19.8 22 28s-.3 10-1.5 10c-1.1 0-1.5-1.2-1.5-4.5V29H8v6.8c0 6.4.2 7 3.3 10 3 2.9 3.9 3.2 9.4 3.2 5.8 0 6.3-.2 9.3-3.6l3.2-3.7 3.7 3.7c3.6 3.5 3.9 3.6 11.6 3.6 7.4 0 8.1-.2 11-3l3.1-3 2.9 3c2.7 2.8 3.4 3 9.9 3 5.8 0 7.4-.4 9.3-2.2l2.3-2.1v2.1c0 2.2.1 2.2 23 2.2h23v-5c0-4.9.1-5 2.9-5 2.1 0 3.7-1 6-3.8l3-3.7.1 3.2c0 2.3-.5 3.3-1.5 3.3-1.2 0-1.5 1.3-1.5 5.5V49h36v-5.5c0-4.2-.3-5.5-1.5-5.5s-1.5-1.5-1.5-7.5.3-7.5 1.5-7.5 1.5-1.3 1.5-5.5V12h-18c-16.1 0-18 .2-18 1.7s-.1 1.5-1.8 0c-1.6-1.5-5-1.7-27.5-1.7H87V16.3l-2.3-2.1c-1.9-1.8-3.5-2.2-9.3-2.2-6.5 0-7.2.2-9.9 3l-2.9 3-3.1-3c-2.9-2.8-3.6-3-11-3-6.7 0-8.2.3-10.2 2.2L36 16.3V7H19v5.5zm14 0c0 .9-.7 1.8-1.5 2.1-1.2.5-1.5 2.9-1.5 13.2 0 12.3-.1 12.7-2.7 15.4-2.1 2.2-3.6 2.8-6.8 2.8-5.7 0-9.5-3.9-9.5-9.8 0-3.5.3-4.2 2-4.2 1.6 0 2 .7 2 3.2 0 3.7 2.5 6.8 5.5 6.8 4 0 4.5-1.8 4.5-14.6 0-9.9-.3-12.3-1.5-12.8-.8-.3-1.5-1.2-1.5-2.1 0-1.2 1.2-1.5 5.5-1.5s5.5.3 5.5 1.5zm24.7 5.7c2.1 2 2.3 3 2.3 12.1 0 9.6-.1 9.9-2.9 12.8-2.6 2.5-3.7 2.9-8.7 2.9-4.9 0-6.2-.4-8.6-2.7-2.7-2.6-2.8-3-2.8-12.7 0-8.9.2-10.2 2.2-12.3 1.8-2 3.1-2.3 9.2-2.3 5.8 0 7.4.4 9.3 2.2zm26.2.4c2.8 3.5 2.8 7.4.1 7.4-1.3 0-2-.7-2-1.9 0-2.9-2.5-4.3-7.2-3.9-3.9.3-4.3.6-4.6 3.1-.5 4.1 1.3 5.7 6.6 5.7 6 0 9.2 2.8 9.2 8 0 2.7-.7 4.3-2.8 6.3-4 3.9-11.4 3.9-15.4 0C64.4 40 63.9 35 67 35c1.5 0 2 .7 2 2.5 0 4.3 7.5 6.2 11.2 2.8 4-3.6 1.5-7.3-5.1-7.3-8.6 0-13.4-8.9-7.9-14.7 1.8-1.9 3.2-2.3 8.4-2.3 5.6 0 6.5.3 8.3 2.6zm29.1 2c0 3.8-.3 4.5-1.7 4.2-1.1-.2-1.9-1.4-2.1-3.1-.3-2.5-.6-2.7-5.3-2.7H99v10h5c3.8 0 5 .4 5 1.5s-1.2 1.5-5 1.5h-5v10h5c4.9 0 5-.1 5-3 0-2.3.4-3 2-3 1.7 0 2 .7 2 5v5h-11c-10.3 0-11-.1-11-2 0-1.1.7-2 1.5-2 1.2 0 1.5-1.9 1.5-11.5S93.7 19 92.5 19c-.8 0-1.5-.7-1.5-1.5 0-1.2 1.9-1.5 11-1.5h11v4.6zm26.7-2.4c3.1 2.9 3.2 9.8.2 13.9-1.9 2.6-2.6 2.9-8 2.9H126v3.5c0 2.9.4 3.5 2.1 3.5 1.4 0 1.9.5 1.7 1.7-.3 1.5-1.4 1.8-5.8 1.8s-5.5-.3-5.8-1.8c-.2-1.2.3-1.7 1.7-1.7 2 0 2.1-.5 2.1-11.5 0-10.8-.1-11.5-2-11.5-1.1 0-2-.7-2-1.5 0-1.2 1.7-1.5 9.7-1.5 8.6 0 9.9.3 12 2.2zm16.3-.7c0 .8-.7 1.5-1.5 1.5-1.1 0-1.5 1.2-1.5 5v5h14v-5c0-3.8-.4-5-1.5-5-.8 0-1.5-.7-1.5-1.5 0-1.2 1.3-1.5 5.5-1.5s5.5.3 5.5 1.5c0 .8-.9 1.5-2 1.5-1.9 0-2 .7-2 11.5 0 11 .1 11.5 2.1 11.5 1.4 0 1.9.5 1.7 1.7-.2 1.4-1.5 1.9-5.6 2.1-4.7.3-5.2.1-5.2-1.7 0-1.2.7-2.1 1.5-2.1 1.1 0 1.5-1.2 1.5-5v-5h-14v5c0 3.8.4 5 1.5 5 .8 0 1.5.9 1.5 2.1 0 1.8-.5 2-5.2 1.7-4.1-.2-5.4-.7-5.6-2.1-.2-1 .3-1.7 1.2-1.7 1.4 0 1.6-1.7 1.6-11.5 0-9.6-.3-11.5-1.5-11.5-.8 0-1.5-.7-1.5-1.5 0-1.2 1.3-1.5 5.5-1.5s5.5.3 5.5 1.5zM78 24c0 .5-1.1 1-2.5 1s-2.5-.5-2.5-1c0-.6 1.1-1 2.5-1s2.5.4 2.5 1zm12.9 6.7c-.1 4.3-.4 6.4-.6 4.8-.3-1.7-.8-3.3-1.2-3.7-.9-1-.1-8.8 1-8.8.5 0 .8 3.5.8 7.7zM106 24c0 .5-.9 1-2 1s-2-.5-2-1c0-.6.9-1 2-1s2 .4 2 1zm11.9 6.7c-.1 4.3-.4 6.6-.6 5-.3-1.9-1-2.7-2.4-2.7s-1.9-.7-1.9-2.5.5-2.5 2-2.5 2-.7 2-2.5c0-1.4.2-2.5.5-2.5s.4 3.5.4 7.7zM163 24c0 .5-1.3 1-3 1-1.6 0-3-.5-3-1 0-.6 1.4-1 3-1 1.7 0 3 .4 3 1zM78 37c0 .5-1.3 1-3 1-1.6 0-3-.5-3-1 0-.6 1.4-1 3-1 1.7 0 3 .4 3 1zm28 0c0 .5-.9 1-2 1s-2-.5-2-1c0-.6.9-1 2-1s2 .4 2 1zm57 0c0 .5-1.3 1-3 1-1.6 0-3-.5-3-1 0-.6 1.4-1 3-1 1.7 0 3 .4 3 1z"
+      d="M19 11c0 2.3.4 3 2 3 1.9 0 2 .7 2 11.3 0 11.9-.6 13.7-4.5 13.7-3 0-4.5-1.9-4.5-5.7 0-3-.3-3.3-3-3.3-2.9 0-3 .1-3 4.7 0 3.7.6 5.3 2.7 7.5C13 44.6 14.1 45 19 45c5.1 0 5.8-.3 7.8-3 2-2.6 2.2-4.1 2.2-15.5 0-10.4.2-12.5 1.5-12.5 1 0 1.5-1 1.5-3V8H19v3zM37.1 15.5c-3 3.5-3.1 3.9-3.1 13.2 0 8.8.2 9.9 2.6 13 2.5 3.2 2.7 3.3 10.1 3.3 7 0 8-.2 10.5-2.7 2.7-2.6 2.8-3 2.8-13 0-9.8-.1-10.6-2.7-13.9-2.7-3.3-3-3.4-9.9-3.4-7 0-7.3.1-10.3 3.5zM52.5 29v9.5l-4.3.3c-6.4.5-7.2-.7-7.2-10.3 0-4.5.3-8.5.7-8.9.4-.4 3-.6 5.8-.4l5 .3V29zM66.2 15.3c-4.2 4.4-4.4 9.8-.5 13.9l2.7 2.8h-2.7c-2.5 0-2.7.3-2.7 3.8 0 2.9.7 4.5 2.8 6.5 2.5 2.4 3.6 2.7 9.2 2.7s6.7-.3 9.2-2.7c3.8-3.7 3.9-9.9.2-13.4l-2.6-2.4 2.4-.5c3.8-.9 3.6-6-.2-10.4-3-3.3-3.6-3.6-8.9-3.6-5.1 0-6 .3-8.9 3.3zm13.6 7.5.3 3.3-4.8-.3c-3.7-.2-4.9-.7-5.1-2.1-.7-3.7.6-4.8 5-4.5 4.1.3 4.3.5 4.6 3.6zm-.3 12.7c0 2.8-.3 3-3.8 3.3-4.5.4-6.7-1.1-6.7-4.5 0-2.3.3-2.4 5.3-2.1 5 .3 5.2.4 5.2 3.3zM90 15.5c0 2.4.4 3.5 1.5 3.5 1.2 0 1.5 1.8 1.5 10s-.3 10-1.5 10c-1 0-1.5 1-1.5 3v3h25V34h-3.5c-3.1 0-3.5.3-3.5 2.5s-.4 2.5-4 2.5c-3.9 0-4-.1-4-3.5V32h5.5c5.4 0 5.5 0 5.5-3s-.1-3-5.5-3H100v-3.5c0-3.4.1-3.5 4-3.5 3.6 0 4 .2 4 2.5 0 2.2.4 2.5 3.5 2.5h3.6l-.3-5.8-.3-5.7-12.2-.3-12.3-.3v3.6zM118 15.5c0 2.4.5 3.5 1.5 3.5 1.2 0 1.5 1.8 1.5 10s-.3 10-1.5 10c-1 0-1.5 1-1.5 3v3h14v-3c0-2.3-.4-3-2-3-1.3 0-2-.7-2-2 0-1.7.7-2 5.3-2 4.5 0 5.7-.4 8-2.8 3.9-4 3.9-11.9 0-16.8l-2.8-3.4H118v3.5zm18.8 7.9c.4 4.4-.7 5.6-5.5 5.6-3.3 0-3.3 0-3.3-5.1v-5l4.3.3c4.1.3 4.2.4 4.5 4.2z"
     />
     <path
       fill="white"
-      d="M42.2 21.2c-.7.7-1.2 4.2-1.2 9.3 0 9.6 1.3 11.5 7.7 11.5 5.8 0 6.4-1.3 6.1-12.1l-.3-9.4-5.5-.3c-3.4-.2-6 .2-6.8 1zm9.8 9.3V38h-7V23h7v7.5zM126 25.5V31h4.8c5.6 0 7.2-1.2 7.2-5.5s-1.6-5.5-7.2-5.5H126v5.5zm8 0c0 1.8-.5 2.5-2 2.5s-2-.7-2-2.5.5-2.5 2-2.5 2 .7 2 2.5zM21 59.5c0 4.2.3 5.5 1.5 5.5S24 66.8 24 75s-.3 10-1.5 10-1.5 1.3-1.5 5.5V96h9c8.1 0 9-.2 9-1.9 0-1.5.8-1.8 4-1.8s4 .3 4 1.8c0 1.7.9 1.9 9 1.9h9v-2.8l.1-2.7 2.4 2.7c2.2 2.6 2.9 2.8 10.2 2.8 7.8 0 7.9 0 11.6-3.8 3.4-3.5 3.6-4.1 3.9-11.2l.3-7.4 3.9 11.2c3.5 9.7 4.3 11.2 6.2 11.2 1.2 0 2.5.6 2.7 1.2.4 1 .8 1 1.4.2.4-.7 1.7-1.3 2.8-1.5 1.6-.3 3-3.2 6.5-13.1 2.5-7.1 5.1-12.8 5.8-12.8.8 0 1.2 2.1 1.2 7.4 0 6.5-.2 7.4-2 7.9-1.7.4-2 1.4-2 5.6V96h46v-5.4l2.8 2.7c2.5 2.4 3.6 2.7 9.8 2.7 6.8 0 7.1-.1 10.7-3.8 3.2-3.2 3.7-4.4 3.7-8.3 0-2.5-.3-4.9-.7-5.2-.3-.4-.5-3.5-.4-6.8.2-5.7 0-6.4-3-9.5-3.1-3.2-3.7-3.4-10.3-3.4-6.2 0-7.3.3-9.8 2.7l-2.8 2.7V59h-58v4.9c0 2.8.4 5.2.9 5.5 1 .7-.8 7.6-2 7.6-1.3 0-2.2-5.4-2-11.8l.3-6.2h-8.5c-5.9 0-8.7.4-9 1.2-.4 1-.6 1-.6 0-.1-.9-2.6-1.2-9-1.2-8.5 0-9.1.1-11 2.6L65 64.3V54H49.1l-2.7 8.5C44.8 67.2 43.3 71 43 71c-.3 0-1.8-3.8-3.4-8.5L36.9 54H21v5.5zm17.5 10.7C40.7 76.7 42.7 82 43 82c.3 0 2.4-5.4 4.5-12l3.9-12h4.8c4.1 0 4.8.3 4.8 2 0 1.1-.7 2-1.5 2-1.3 0-1.5 2.2-1.5 13.5S58.2 89 59.5 89c.8 0 1.5.9 1.5 2.1 0 1.8-.5 2-5.2 1.7-4.1-.2-5.4-.7-5.6-2.1-.2-1.2.3-1.7 1.7-1.7 2 0 2.1-.5 2.1-12 0-6.6-.2-12-.4-12-.3 0-1 1.7-1.7 3.7C47.2 83.4 44.7 89 43 89c-1.9 0-3.2-2.8-7.9-17.3-1.3-3.7-2.5-6.7-2.7-6.7-.2 0-.4 5.4-.4 12 0 11.5.1 12 2.1 12 1.4 0 1.9.5 1.7 1.8-.2 1.3-1.5 1.8-5.6 2-4.7.3-5.2.1-5.2-1.7 0-1.2.7-2.1 1.5-2.1 1.3 0 1.5-2.2 1.5-13.5S27.8 62 26.5 62c-.8 0-1.5-.9-1.5-2.1 0-1.8.5-2 4.8-1.7l4.7.3 4 11.7zm48.2-5c2.1 2 2.3 3 2.3 12.3 0 9.8-.1 10.2-2.8 12.8-4.1 3.9-12.3 3.9-16.4 0-2.7-2.6-2.8-3-2.8-12.5 0-13.6.8-14.8 10.3-14.8 5.9 0 7.5.4 9.4 2.2zm15.3-.7c0 .8-.7 1.5-1.6 1.5-1.8 0-1.8 0 3.2 13.5 3.6 10.1 2.8 10.4 7.7-3 1.4-3.9 2.8-7.8 3.2-8.8.5-1.2.2-1.7-.9-1.7-.9 0-1.6-.7-1.6-1.5 0-1.1 1.2-1.5 5-1.5s5 .4 5 1.5c0 .8-.6 1.5-1.4 1.5-.8 0-3 4.6-5.6 12.2-4.9 13.9-5.4 14.8-8.5 14.8-2.4 0-2.3.2-8.6-17.3-1.7-4.9-3.8-9.1-4.5-9.4-.8-.3-1.4-1.1-1.4-1.9 0-1 1.4-1.4 5-1.4 3.8 0 5 .4 5 1.5zm34 0c0 .8-.9 1.5-2 1.5-1.9 0-2 .7-2 11.5 0 11 .1 11.5 2.1 11.5 1.4 0 1.9.5 1.7 1.8-.2 1.3-1.5 1.8-6.1 2-5.2.3-5.7.1-5.7-1.8 0-1.3.7-2 2-2 1.9 0 2-.7 2-11.5s-.1-11.5-2-11.5c-1.1 0-2-.7-2-1.5 0-1.2 1.3-1.5 6-1.5s6 .3 6 1.5zm26.8 2.7c-.2 3.1-.8 4.4-2 4.6-1.4.3-1.8-.3-1.8-2.7 0-3.1 0-3.1-5.6-3.1h-5.5l.3 4.7.3 4.8 5.3.3c3.7.2 5.2.7 5.2 1.8 0 1-1.4 1.4-5.5 1.4H148v10h5.5c5.4 0 5.5 0 5.5-3 0-2.4.4-3 2.1-3 1.8 0 2 .5 1.7 4.7l-.3 4.8-10.7.3c-10.3.3-10.8.2-10.8-1.8 0-1.1.7-2 1.5-2 1.2 0 1.5-1.9 1.5-11.5s-.3-11.5-1.5-11.5c-.8 0-1.5-.7-1.5-1.5 0-1.2 1.9-1.5 11.1-1.5h11l-.3 4.2zm24.1-1.6c2.8 3.5 2.8 7.4.1 7.4-1.3 0-2-.7-2-1.9 0-2.9-2.5-4.3-7.2-3.9-3.9.3-4.3.6-4.6 3.1-.5 4.1 1.3 5.7 6.6 5.7 6 0 9.2 2.8 9.2 8 0 5.5-4.1 9-10.5 9-3.9 0-5.4-.5-7.7-2.7-3.4-3.3-3.9-8.3-.8-8.3 1.2 0 2 .7 2 1.7 0 4.6 7.4 7.1 11.2 3.6 4-3.6 1.5-7.3-5.1-7.3-8.6 0-13.4-8.9-7.9-14.7 1.8-1.9 3.2-2.3 8.4-2.3 5.6 0 6.5.3 8.3 2.6zM63.2 72.2c-.4 12.5-.9 13.7-1 3-.1-5.6.1-10.2.6-10.2.4 0 .6 3.3.4 7.2zm76.8 5.3c0 6.8-.2 7.5-2 7.5s-2-.7-2-7.5.2-7.5 2-7.5 2 .7 2 7.5zm15-6.5c0 .5-.7 1-1.5 1s-1.5-.5-1.5-1c0-.6.7-1 1.5-1s1.5.4 1.5 1zm26 0c0 .5-1.1 1-2.5 1s-2.5-.5-2.5-1c0-.6 1.1-1 2.5-1s2.5.4 2.5 1zm-15.7 5.3c.2.6-.2 1.8-.9 2.5-1.2 1.1-1.4.8-1.4-1.3 0-2.6 1.4-3.4 2.3-1.2zM155 84c0 .5-.7 1-1.5 1s-1.5-.5-1.5-1c0-.6.7-1 1.5-1s1.5.4 1.5 1zm26 0c0 .5-1.1 1-2.4 1-1.4 0-2.8-.5-3.1-1-.4-.6.7-1 2.4-1s3.1.4 3.1 1z"
-    />
-    <path
-      fill="white"
-      d="M71.7 67.6c-.4.4-.7 4.9-.7 10C71 88 71.6 89 78.3 89c5.7 0 6.8-2.3 6.5-12.6l-.3-8.9-6-.3c-3.4-.2-6.4 0-6.8.4zm9.3 9.9V85h-6V70h6v7.5z"
+      d="M145.7 12.7c-1.5 1.5-.7 5.5 1.3 6 1.8.5 2 1.4 2 10.3s-.2 9.8-2 10.3c-1.4.3-2 1.4-2 3.1 0 2.6.1 2.6 7 2.6h7v-3c0-2-.5-3-1.5-3s-1.5-1.1-1.5-3.5V32h11v3.5c0 2.4-.5 3.5-1.5 3.5s-1.5 1-1.5 3v3h7c6.9 0 7 0 7-2.6 0-1.7-.6-2.8-2-3.1-1.8-.5-2-1.4-2-10.3 0-9 .2-9.8 2.1-10.3 1.6-.4 2-1.2 1.7-3.4-.3-2.7-.5-2.8-7-3.1l-6.8-.3V15c0 1.8.6 3.3 1.5 3.6.9.4 1.5 1.9 1.5 4V26h-11v-3.4c0-2.1.6-3.6 1.5-4 .9-.3 1.5-1.8 1.5-3.6v-3h-6.3c-3.5 0-6.7.3-7 .7zM22 59c0 2 .5 3 1.5 3 1.3 0 1.5 2.1 1.5 12.5S24.8 87 23.5 87c-1 0-1.5 1.1-1.5 3.5V94h13v-3.5c0-2.4-.5-3.5-1.5-3.5-1.2 0-1.5-1.5-1.4-8.3l.1-8.2 3.2 9.5c3.1 9 3.4 9.5 6.2 9.8 3 .3 3.1.1 6.1-8.5l3.2-8.8v7.2c.1 6-.2 7.3-1.5 7.3-1.1 0-1.4.8-1.2 3.2l.3 3.3H55c6.4 0 6.5 0 6.8-2.8.3-2.2-.1-3-1.7-3.4-2-.5-2.1-1.2-2.1-12.9 0-11.7.1-12.4 2-12.4 1.6 0 2-.7 2-3v-3.1l-6.1.3-6.1.3L46 67.3c-2.1 6-4 11-4.2 11.3-.2.2-2.1-4.8-4.2-11.1L33.9 56h-6C22 56 22 56 22 59zM117.3 60.8c-6 .3-6.3.4-6.3 2.8 0 1.3.6 2.7 1.4 3 1 .4.7 2.3-1.8 8.9-1.6 4.7-3.3 8.1-3.7 7.7-1.7-1.7-6-15.7-5-16 .6-.2 1.1-1.7 1.1-3.3 0-2.9 0-2.9-6-2.9h-6v3c0 1.9.5 3 1.4 3 .8 0 3.2 5 6.2 13.2l4.9 13.3 3.6.3 3.6.3 4.7-12.6c2.5-7 4.6-13.1 4.6-13.6s1.6-.9 3.5-.9h3.5v10c0 8.2-.3 10-1.5 10-1 0-1.5 1.1-1.5 3.5V94h14v-3.5c0-2.8-.4-3.5-2-3.5-1.9 0-2-.7-2-10s.1-10 2-10c1.5 0 2-.7 2-2.9v-2.9l-7.2-.4c-4-.2-10.1-.2-13.5 0zM67.2 64c-1.9 2.6-2.2 4.2-2.2 13.2 0 10.1.1 10.4 2.9 13.6 2.8 3 3.4 3.2 9.6 3.2 6.2 0 6.9-.2 9.6-3.1 2.8-3 2.9-3.4 2.9-13.9 0-9.3-.3-11.1-2.1-13.4C86 61.2 85.3 61 77.6 61c-8 0-8.3.1-10.4 3zm15.3 13v9.5h-10l-.3-8.4c-.2-4.6 0-9.1.2-9.8.4-1 2-1.3 5.3-1.1l4.8.3V77zM141 64c0 2 .5 3 1.5 3 1.2 0 1.5 1.8 1.5 10s-.3 10-1.5 10c-1.1 0-1.5 1.1-1.5 3.5v3.6l12.3-.3 12.2-.3.3-5.7.3-5.8h-3.6c-3.1 0-3.5.3-3.5 2.5s-.4 2.5-4 2.5c-3.9 0-4-.1-4-3.5V80h5.5c5.4 0 5.5 0 5.5-3s-.1-3-5.5-3H151v-3.5c0-3.4.1-3.5 4-3.5 3.6 0 4 .2 4 2.5 0 2.2.4 2.5 3.5 2.5h3.5V61h-25v3zM171.3 64c-3 3.7-3.1 10-.3 12.5 2.7 2.4 2.5 3.5-.5 3.5-3.8 0-3.5 5.8.6 10.5 2.9 3.3 3.4 3.5 9.4 3.5s6.5-.2 9.4-3.5c4.2-4.8 4.2-8.8-.1-13.2l-3.2-3.3h2.7c2.5 0 2.7-.3 2.7-3.8 0-2.9-.7-4.5-2.8-6.5-2.5-2.4-3.6-2.7-9.3-2.6-5.8 0-6.6.3-8.6 2.9zm13.5 4.2c.7.7 1.2 2.3 1.2 3.6 0 2.1-.3 2.3-4.7 2-4.9-.3-6.3-1.8-4.9-5.4.7-1.7 6.7-1.9 8.4-.2zm.7 15.8c0 2.3-.3 2.5-5 2.5s-5-.1-5.3-2.8c-.3-2.8-.3-2.8 5-2.5 4.9.3 5.3.5 5.3 2.8z"
     />
   </svg>
 );
