+++ src/components/Detail/Detail.tsx
@@ -1,10 +1,10 @@
 import React, { useState, useEffect } from "react";
-import styled from "styled-components";
 import axios from "axios";
-import { Button, Modal, FixedHeadingStyles, BodyStyles } from "joseph-ui-kit";
-import { Trailer } from "./Youtube";
-import SkeletonDetail from "./SkeletonDetail";
-import ScoreGraph from "./ScoreGraph";
+import { Button, Modal } from "joseph-ui-kit";
+import { Trailer } from "../Youtube/Youtube";
+import SkeletonDetail from "../SkeletonDetail/SkeletonDetail";
+import ScoreGraph from "../ScoreGraph/ScoreGraph";
+import * as Styled from "./Detail.styles";
 
 interface DetailProps {
   movie_id: number;
@@ -48,23 +48,23 @@ const Detail = ({ movie_id }: DetailProps) => {
   return loading ? (
     <SkeletonDetail />
   ) : (
-    <Container>
+    <Styled.Container>
       {detail.poster_path === null ? (
-        <NullImage />
+        <Styled.NullImage />
       ) : (
-        <MovieImage
+        <Styled.MovieImage
           src={`https://image.tmdb.org/t/p/w200/${detail?.poster_path}`}
           alt="movie poster image"
         />
       )}
-      <Description>
-        <Title>
+      <Styled.Description>
+        <Styled.Title>
           {detail.original_title +
             (detail.release_date
               ? ` (${detail.release_date.slice(0, 4)})`
               : "")}
-        </Title>
-        <InfoBox>
+        </Styled.Title>
+        <Styled.InfoBox>
           <ScoreGraph vote_average={detail?.vote_average} />
           {videoId ? (
             <Button
@@ -74,8 +74,8 @@ const Detail = ({ movie_id }: DetailProps) => {
               onClick={openModal}
             />
           ) : null}
-        </InfoBox>
-        <GenreContainer>
+        </Styled.InfoBox>
+        <Styled.GenreContainer>
           <div>
             {detail.release_date}
             {detail.production_countries.map(
@@ -94,13 +94,13 @@ const Detail = ({ movie_id }: DetailProps) => {
                 genre.name + (index !== detail.genres.length - 1 ? ", " : "")
             )}
           </div>
-        </GenreContainer>
-        <Tagline>
+        </Styled.GenreContainer>
+        <Styled.Tagline>
           <em>{detail.tagline}</em>
-        </Tagline>
-        <Overview>{detail.overview ? "개요" : null}</Overview>
-        <Paragraph>{detail.overview}</Paragraph>
-      </Description>
+        </Styled.Tagline>
+        <Styled.Overview>{detail.overview ? "개요" : null}</Styled.Overview>
+        <Styled.Paragraph>{detail.overview}</Styled.Paragraph>
+      </Styled.Description>
 
       {ModalOpen ? (
         <Modal
@@ -115,98 +115,8 @@ const Detail = ({ movie_id }: DetailProps) => {
           <Trailer videoId={videoId} />
         </Modal>
       ) : null}
-    </Container>
+    </Styled.Container>
   );
 };
 
 export default Detail;
-
-const Container = styled.div`
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  margin-top: 20px;
-  @media screen and (min-width: 800px) {
-    flex-direction: row;
-    align-items: flex-start;
-    margin-top: 50px;
-    margin-bottom: 50px;
-  }
-`;
-
-const MovieImage = styled.img`
-  width: 200px;
-  height: 300px;
-  object-fit: cover;
-`;
-
-const NullImage = styled.div`
-  min-width: 200px;
-  min-height: 300px;
-  background-image: url("https://www.themoviedb.org/assets/2/v4/glyphicons/basic/glyphicons-basic-38-picture-grey-c2ebdbb057f2a7614185931650f8cee23fa137b93812ccb132b9df511df1cfac.svg");
-  background-position: center;
-  background-repeat: no-repeat;
-  background-size: 80%;
-  background-color: #dbdbdb;
-  object-fit: contain;
-  cursor: pointer;
-  @media screen and (max-width: 320px) {
-    width: 120px;
-    height: 180px;
-  }
-`;
-
-const Description = styled.div`
-  margin: 20px 0px;
-
-  @media screen and (min-width: 800px) {
-    margin: 0;
-    margin-left: 50px;
-  }
-`;
-
-const DefaultStyle = `
-  text-align: center;
-  margin: 10px;
-  @media screen and (min-width: 800px) {
-    text-align: left;
-  }
-`;
-
-const Title = styled.div`
-  ${DefaultStyle}
-  ${FixedHeadingStyles.external.heading04}
-`;
-
-const InfoBox = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  ${DefaultStyle}
-  @media screen and (min-width: 800px) {
-    justify-content: flex-start;
-  }
-`;
-
-const GenreContainer = styled.div`
-  ${DefaultStyle}
-  ${BodyStyles.external.bodyCompact02}
-`;
-
-const Tagline = styled.div`
-  ${DefaultStyle}
-  ${FixedHeadingStyles.external.heading02}
-  color: gray;
-  font-style: italic;
-`;
-const Overview = styled.div`
-  ${DefaultStyle}
-  ${FixedHeadingStyles.external.heading03}
-  text-align: left;
-`;
-
-const Paragraph = styled.p`
-  ${DefaultStyle}
-  ${BodyStyles.external.body02}
-`;
