+++ src/components/Footer/Footer.styles.tsx
@@ -1,0 +1,21 @@
+import styled from "styled-components";
+
+export const Container = styled.footer`
+  width: 100%;
+  height: 60px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  background-color: #376188;
+`;
+
+export const LogoBox = styled.div`
+  width: 130px;
+  display: flex;
+  justify-content: space-between;
+`;
+
+export const GitHubAnchor = styled.a`
+  text-decoration: none;
+  color: white;
+`;
