+++ src/components/NowPlayingTabPanel/NowPlayingTabPanel.tsx
@@ -1,14 +1,14 @@
 import React, { useState, useEffect, useRef } from "react";
-import styled from "styled-components";
 import axios from "axios";
 import { Button } from "joseph-ui-kit";
-import Contents from "./Contents";
-import SkeletonContents from "./SkeletonContents";
-import SortAccordion from "./SortAccordion";
+import MovieContents from "../MovieContents/MovieContents";
+import SkeletonContents from "../SkeletonContents/SkeletonContents";
+import SortAccordion from "../SortAccordion/SortAccordion";
+import * as Styled from "./NowPlayingTabPanel.styles";
 
-const PopularTopPanel = () => {
+const NowPlayingTopPanel = () => {
   const [loading, setLoading] = useState(true);
-  const [popular, setPopular] = useState([]);
+  const [nowPlaying, setNowPlaying] = useState([]);
   const [page, setPage] = useState(1);
 
   const [sortedMovies, setSortedMovies] = useState([]);
@@ -33,16 +33,15 @@ const PopularTopPanel = () => {
     setLoading(true);
     axios
       .get(
-        `https://api.themoviedb.org/3/movie/popular?api_key=${process.env.REACT_APP_API_KEY}&language=ko&page=${page}`
+        `https://api.themoviedb.org/3/movie/now_playing?api_key=${process.env.REACT_APP_API_KEY}&language=ko&page=${page}`
       )
-      .then((res) => {
-        setPopular((popular) => popular.concat(res.data.results));
-      })
+      .then((res) =>
+        setNowPlaying((nowPlaying) => nowPlaying.concat(res.data.results))
+      )
       .then(() => setLoading(false))
       .catch((err) => console.log(err));
   }, [page]);
 
-  // 첫 랜더링은 무시 -> 이후 정렬 DropDown을 통해 값(selected)이 변경되면 실행
   useEffect(() => {
     if (didMount.current) {
       setLoading(true);
@@ -89,32 +88,24 @@ const PopularTopPanel = () => {
   ) : sortedMovies.length === 0 ? (
     <>
       <SortAccordion setSelected={setSelected} />
-      <Contents data={popular} />
-      {popular.length >= 20 * page ? (
-        <ButtonContainer>
+      <MovieContents data={nowPlaying} />
+      {nowPlaying.length >= 20 * page ? (
+        <Styled.ButtonContainer>
           <Button name="더보기" padding="10px 70px" onClick={morePage} />
-        </ButtonContainer>
+        </Styled.ButtonContainer>
       ) : null}
     </>
   ) : (
     <>
       <SortAccordion setSelected={setSelected} />
-      <Contents data={sortedMovies} />
+      <MovieContents data={sortedMovies} />
       {sortedMovies.length >= 20 * sortedPage ? (
-        <ButtonContainer>
+        <Styled.ButtonContainer>
           <Button name="더보기" padding="10px 70px" onClick={moreSortedPage} />
-        </ButtonContainer>
+        </Styled.ButtonContainer>
       ) : null}
     </>
   );
 };
 
-export default PopularTopPanel;
-
-const ButtonContainer = styled.div`
-  width: 100%;
-  height: 100px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-`;
+export default NowPlayingTopPanel;
