+++ src/components/PopularTabPanel/PopularTabPanel.styles.tsx
@@ -1,0 +1,9 @@
+import styled from "styled-components";
+
+export const ButtonContainer = styled.div`
+  width: 100%;
+  height: 100px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+`;
