+++ src/components/Youtube/Youtube.styles.tsx
@@ -1,0 +1,9 @@
+import styled from "styled-components";
+
+export const VideoBox = styled.div`
+  width: 100%;
+  height: 100%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+`;
